backend/requirements.txt
fastapi
uvicorn[standard]
mlb-statsapi
python-dotenv
